refactor(sales):move methods to different file
