+++ .gitignore
@@ -74,3 +74,4 @@ Makefile
 # local-only docs
 docs/migration-runbook.md
 backup_prod_20260521.sql
+docs/flow_start-sales-process.md
